% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal.R
\name{.add_missing}
\alias{.add_missing}
\title{add missing}
\usage{
.add_missing(df, GroupNames, prop_missing)
}
\arguments{
\item{df}{A data frame containing the simulated data with columns 'GroupName', 'y1', and 'y2'.}

\item{GroupNames}{A character vector of length 2 specifying the group names corresponding to the two groups in the data frame.}

\item{prop_missing}{A numeric vector of length 2 specifying the proportion of missing values
for each group. The first element corresponds to the first group in GroupNames, and the second element corresponds to the second group.}
}
\value{
A modified data frame with missing values added to 'y1' and 'y2' according to the specified proportions, and new ordinal columns 'Ord_1' and 'Ord_2' added based on the cutpoints applied to 'y1' and 'y2 respectively.
}
\description{
Internal function to add missingness to the simulated data frame.
}
\details{
Apply random missingness by GroupName (not by R value, which breaks when groups share the same relatedness),
}
