% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal.R
\name{.add_missing_and_ordinal}
\alias{.add_missing_and_ordinal}
\title{add_missing_and_ordinal}
\usage{
.add_missing_and_ordinal(df, GroupNames, prop_missing)
}
\arguments{
\item{df}{A data frame containing the simulated data with columns 'GroupName', 'y1', and 'y2'.}

\item{GroupNames}{A character vector of length 2 specifying the group names corresponding to the two groups in the data frame.}

\item{prop_missing}{A numeric vector of length 2 specifying the proportion of missing values
for each group. The first element corresponds to the first group in GroupNames, and the second element corresponds to the second group.}
}
\value{
A modified data frame with missing values added to 'y1' and 'y2' according to the specified proportions, and new ordinal columns 'Ord_1' and 'Ord_2' added based on the cutpoints applied to 'y1' and 'y2 respectively.
}
\description{
add_missing_and_ordinal
}
\details{
Applies random missingness to the 'y1' and 'y2' columns of the input data frame based on the specified proportions for each group, and
then compute 4-category ordinal columns.
Cutpoints: (-Inf, -2] = 1, (-2, -1] = 2, (-1, 1) = 3, [1, Inf) = 4.
NA y values produce NA ordinal scores; missingness is cascaded from y1 to y2.
}
