% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harvest.R
\name{harvest}
\alias{harvest}
\title{Harvest results from MCMCglmm analyses}
\usage{
harvest(results_fit, type)
}
\arguments{
\item{results_fit}{A list of model fitting results generated from \code{Sim_Fit2}. This list should have two sub-lists: one for "Interval" model results and one for "Ordinal" model results. Each sub-list should contain the fitting results for each iteration (e.g., "Iteration1", "Iteration2", etc.).}

\item{type}{A character string specifying the type of model results to extract. Should be
either "Interval" for the interval model results or "Ordinal" for the ordinal model results.}
}
\value{
Returns a \code{data.frame} with the following columns:
\item{Iteration}{The iteration name (e.g., "Iteration1", "
Iteration2", etc.)}
\item{Analysis}{The type of analysis ("Interval" or "Ordinal")}
\item{VA11}{The estimated additive genetic variance component (A) for the
first group of kin pairs}
\item{VC11}{The estimated common environmental variance component (C) for the
first group of kin pairs}
\item{VE11}{The estimated unique environmental variance component (E) for the
first group of kin pairs}
}
\description{
A function to extract the A, C, E estimates from the model
fitting results of the MCMCglmm analyses. This function is designed to work with the output of \code{Sim_Fit2}, which includes both interval and ordinal model fitting results.
}
