% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal.R
\name{rmvn}
\alias{rmvn}
\title{rmvn}
\usage{
rmvn(n, sigma)
}
\arguments{
\item{n}{Sample Size}

\item{sigma}{Covariance matrix}
}
\value{
Generates multivariate normal data from a covariance matrix (\code{sigma}) of length \code{n}
}
\description{
rmvn
}
\keyword{internal}
